+++ sharedbuffers/mapped_struct.py
@@ -13,7 +13,8 @@ import sys
 import xxhash
 import itertools
 import time
-from datetime import timedelta, datetime
+import zipfile
+from datetime import timedelta, datetime, date
 from decimal import Decimal
 
 try:
@@ -975,6 +976,32 @@ class mapped_datetime(datetime):
             idmap[offs] = rv
         return rv
 
+class mapped_date(date):
+    @classmethod
+    @cython.locals(offs = cython.longlong, implicit_offs = cython.longlong, timestamp = cython.longlong)
+    def pack_into(cls, obj, buf, offs, idmap = None, implicit_offs = 0, packer = struct.Struct('=q')):
+        if idmap is not None:
+            objid = id(obj)
+            idmap[objid] = offs + implicit_offs
+
+        timestamp = int(time.mktime(obj.timetuple()))
+        packer.pack_into(buf, offs, timestamp)
+
+        return offs + packer.size
+
+    @classmethod
+    @cython.locals(offs = cython.longlong, timestamp = cython.longlong)
+    def unpack_from(cls, buf, offs, idmap = None, packer = struct.Struct('=q')):
+        if idmap is not None and offs in idmap:
+            return idmap[offs]
+
+        timestamp, = packer.unpack_from(buf, offs)
+        rv =  date.fromtimestamp(timestamp)
+
+        if idmap is not None:
+            idmap[offs] = rv
+        return rv
+
 class mapped_object(object):
     __slots__ = ('value', 'typecode')
 
@@ -994,21 +1021,23 @@ class mapped_object(object):
         mapped_frozenset : 'Z',
         mapped_tuple : 't',
         mapped_list : 'e',
-        mapped_bytes : 's',
         mapped_unicode : 'u',
         mapped_dict : 'm',
-
-        proxied_list: 'E',
-        proxied_tuple: 'V',
-
-        mapped_datetime : 'z',
+        mapped_bytes : 's',
+        mapped_datetime : 'v',
+        mapped_date : 'V',
         mapped_decimal : 'F',
 
+        proxied_list: 'E',
+        proxied_tuple: 'W',
+
         int : 'q',
+        long : 'q',
         float : 'd',
         str : 's',
         unicode : 'u',
-        datetime : 'z',
+        datetime : 'v',
+        date : 'V',
         Decimal : 'F',
         cDecimal : 'F',
     }
@@ -1037,10 +1066,10 @@ class mapped_object(object):
         's' : (mapped_bytes.pack_into, mapped_bytes.unpack_from, mapped_bytes),
         'u' : (mapped_unicode.pack_into, mapped_unicode.unpack_from, mapped_unicode),
         'm' : (mapped_dict.pack_into, mapped_dict.unpack_from, mapped_dict),
-
-        'V' : (proxied_tuple.pack_into, proxied_tuple.unpack_from, proxied_tuple),
+        'W' : (proxied_tuple.pack_into, proxied_tuple.unpack_from, proxied_tuple),
         'E' : (proxied_list.pack_into, proxied_list.unpack_from, proxied_list),
-        'z' : (mapped_datetime.pack_into, mapped_datetime.unpack_from, mapped_datetime),
+        'v' : (mapped_datetime.pack_into, mapped_datetime.unpack_from, mapped_datetime),
+        'V' : (mapped_date.pack_into, mapped_date.unpack_from, mapped_date),
         'F' : (mapped_decimal.pack_into, mapped_decimal.unpack_from, mapped_decimal),
     }
 
@@ -1127,6 +1156,7 @@ VARIABLE_TYPES = {
     bytes : mapped_bytes,
     object : mapped_object,
     datetime : mapped_datetime,
+    date : mapped_date,
     Decimal : mapped_decimal,
     cDecimal : mapped_decimal,
 }
@@ -1536,6 +1566,10 @@ class DatetimeBufferProxyProperty(GenericBufferProxyProperty):
     typ = mapped_datetime
 
 @cython.cclass
+class DateBufferProxyProperty(GenericBufferProxyProperty):
+    typ = mapped_date
+
+@cython.cclass
 class DecimalBufferProxyProperty(GenericBufferProxyProperty):
     typ = mapped_decimal
 
@@ -1581,16 +1615,19 @@ PROXY_TYPES = {
     mapped_bytes : BytesBufferProxyProperty,
     mapped_object : ObjectBufferProxyProperty,
     mapped_datetime : DatetimeBufferProxyProperty,
+    mapped_date : DateBufferProxyProperty,
     mapped_decimal : DecimalBufferProxyProperty,
 
     proxied_tuple : ProxiedTupleBufferProxyProperty,
     proxied_list : ProxiedListBufferProxyProperty,
 
     int : LongBufferProxyProperty,
+    long : LongBufferProxyProperty,
     float : DoubleBufferProxyProperty,
     str : BytesBufferProxyProperty,
     unicode : UnicodeBufferProxyProperty,
     datetime : DatetimeBufferProxyProperty,
+    date : DateBufferProxyProperty,
     Decimal : DecimalBufferProxyProperty,
     cDecimal : DecimalBufferProxyProperty,
 }
@@ -1917,7 +1954,18 @@ class Schema(object):
                     val_offs = idmap_get(val_id)
                     if val_offs is None:
                         idmap[val_id] = ival_offs = offs + implicit_offs
-                        offs = slot_types[slot].pack_into(val, buf, offs, idmap, implicit_offs)
+                        try:
+                            offs = slot_types[slot].pack_into(val, buf, offs, idmap, implicit_offs)
+                        except Exception as e:
+                            try:
+                                # Add some context. It may not work with all exception types, hence the fallback
+                                e = type(e)("%s packing attribute %s=%r of type %r" % (
+                                    e, slot, val, type(obj).__name__))
+                            except:
+                                pass
+                            else:
+                                raise e
+                            raise
                         padding = (offs + alignment - 1) / alignment * alignment - offs
                         offs += padding
                     else:
@@ -1938,7 +1986,15 @@ class Schema(object):
             packer, padding = self.get_packer(obj)
         baseoffs = offs
         packable, offs = self.get_packable(packer, padding, obj, offs, buf, idmap, implicit_offs)
-        packer.pack_into(buf, baseoffs, *packable)
+        try:
+            packer.pack_into(buf, baseoffs, *packable)
+        except struct.error as e:
+            raise struct.error("%s packing %r with format %r for type %s" % (
+                e,
+                packable,
+                packer.format,
+                type(obj).__name__,
+            ))
         if offs > len(buf):
             raise RuntimeError("Buffer overflow")
         return offs
@@ -2145,7 +2201,34 @@ class mapped_object_with_schema(object):
     def unpack_from(self, buf, offs, idmap = None):
         return self.schema.unpack_from(buf, offs, idmap)
 
-class MappedArrayProxyBase(object):
+@cython.ccall
+def _map_zipfile(cls, fileobj, offset, size):
+    # Open underlying file
+    if fileobj._compress_type != zipfile.ZIP_STORED:
+        raise ValueError("Can only map uncompressed elements of a zip file")
+    if fileobj._decrypter is not None:
+        raise ValueError("Cannot map from an encrypted zip file")
+
+    if size is None:
+        size = fileobj._compress_size - offset
+    else:
+        size = min(size, fileobj._compress_size - offset)
+    offset += fileobj._fileobj.tell()
+
+    return cls.map_file(fileobj._fileobj, offset, size)
+
+class _ZipMapBase(object):
+    @classmethod
+    def map_zipfile(cls, fileobj, offset = 0, size = None):
+        return _map_zipfile(cls, fileobj, offset, size)
+
+@cython.cclass
+class _CZipMapBase(object):
+    @classmethod
+    def map_zipfile(cls, fileobj, offset = 0, size = None):
+        return _map_zipfile(cls, fileobj, offset, size)
+
+class MappedArrayProxyBase(_ZipMapBase):
     _CURRENT_VERSION = 2
     _CURRENT_MINIMUM_READER_VERSION = 2
 
@@ -2315,6 +2398,9 @@ class MappedArrayProxyBase(object):
 
     @classmethod
     def map_file(cls, fileobj, offset = 0, size = None):
+        if isinstance(fileobj, zipfile.ZipExtFile):
+            return cls.map_zipfile(fileobj, offset, size)
+
         fileobj.seek(offset)
         total_size = cls._Header.unpack(fileobj.read(cls._Header.size))[0]
         map_start = offset - offset % mmap.ALLOCATIONGRANULARITY
@@ -3212,7 +3298,7 @@ def _discard_duplicates(apart, struct_dt, discard_duplicate_keys, discard_duplic
     return apart
 
 @cython.cclass
-class NumericIdMapper(object):
+class NumericIdMapper(_CZipMapBase):
     dtype = npuint64
 
     # Num-items, Index-pos
@@ -3634,6 +3720,9 @@ class NumericIdMapper(object):
     @classmethod
     @cython.locals(rv = 'NumericIdMapper')
     def map_file(cls, fileobj, offset = 0, size = None):
+        if isinstance(fileobj, zipfile.ZipExtFile):
+            return cls.map_zipfile(fileobj, offset, size)
+
         map_start = offset - offset % mmap.ALLOCATIONGRANULARITY
         fileobj.seek(map_start)
         buf = mmap.mmap(fileobj.fileno(), 0, access = mmap.ACCESS_READ, offset = map_start)
@@ -3693,7 +3782,7 @@ def safe_utf8(x):
         return x
 
 @cython.cclass
-class StringIdMapper(object):
+class StringIdMapper(_CZipMapBase):
     encode = staticmethod(safe_utf8)
     dtype = npuint64
     xxh = xxhash.xxh64
@@ -4137,6 +4226,9 @@ class StringIdMapper(object):
     @classmethod
     @cython.locals(rv = 'StringIdMapper')
     def map_file(cls, fileobj, offset = 0, size = None):
+        if isinstance(fileobj, zipfile.ZipExtFile):
+            return cls.map_zipfile(fileobj, offset, size)
+
         map_start = offset - offset % mmap.ALLOCATIONGRANULARITY
         fileobj.seek(map_start)
         buf = mmap.mmap(fileobj.fileno(), 0, access = mmap.ACCESS_READ, offset = map_start)
@@ -4683,7 +4775,7 @@ def _iter_key_dump(keys_file):
         except EOFError:
             break
 
-class MappedMappingProxyBase(object):
+class MappedMappingProxyBase(_ZipMapBase):
     # Must subclass to select mapping strategies
 
     # A MappedArrayProxyBase subclass for values
@@ -4783,6 +4875,9 @@ class MappedMappingProxyBase(object):
 
     @classmethod
     def map_file(cls, fileobj, offset = 0, size = None):
+        if isinstance(fileobj, zipfile.ZipExtFile):
+            return cls.map_zipfile(fileobj, offset, size)
+
         # If no size is given, it's the whole file by default
         if size is None:
             fileobj.seek(0, os.SEEK_END)